moved code out of reducer into actions file
